feat: add double, numeric extensions in coremodule
