+++ TurboBudget/Core/Components/CircleCategory.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import CoreModule
 
 struct CircleCategory: View {
