+++ TurboBudget/Epic/Budgets/Views/BudgetRowView.swift
@@ -9,6 +9,7 @@
 import SwiftUI
 import TheoKit
 import DesignSystemModule
+import CoreModule
 
 struct BudgetRowView: View {
