+++ TurboBudget/Epic/Generic/Views/Progress/ProgressBarWithAmount.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import CoreModule
 
 struct ProgressBarWithAmount: View {
