+++ TurboBudget/Epic/Generic/Views/TextFields/CustomTextField.swift
@@ -8,6 +8,7 @@
 import SwiftUI
 import TheoKit
 import DesignSystemModule
+import CoreModule
 
 enum CustomTextFieldStyle {
     case text
