+++ TurboBudget/Epic/Statistics/Views/StatisticsRowView.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import CoreModule
 
 struct StatisticsRowView: View {
